Killdevil edits
Edited fishhook and srafish to work on the kill devil cluster.

Added an extra option to fishhook to take a path to the open ssh key.

Changed the hard coded reference index directory in srafish.pl. Added
an $EXECUTE flag so command can be seen without being executed.
Finally, there are very few non Col-0 datasets and for those that are,
mapping to Col-0 works nearly as well. So instead of rebuilding all of
the indexes of the 19 magic parents (lots of disk space and time),
determine_genotype subroutine now always returns Col-0.