srafish code refactoring
Major edits:

The code is getting long and jungly. The edits in this commit are
trying to make the code simpler and more modular to improve readability
and extensibility.

1. Creating a new subroutine called query_details, which basically
takes a line of the query table and reports back several things about
the query. These include:
- The run ID
- Whether it’s included in a supplied “do not process” list.
- Whether the run has already been processed completely.

2. The already_processed subroutine is overly verbose. Currently, the
code opens the directory, loops through all of the files to see if
quant_bias_corrected.sf exists. Additionally it performs a function it
should not perform, which is to remove the directory if its only
partially processed. The function should do ONLY what its name
suggests, i.e. to answer whether the run’s been processed.

To fix this, I am simply deleting the ‘old’ already_processed and
renaming completion_check to ‘processed_to_completion’, which more
directly achieves our goals.

3. Renamed the %processed hash to ‘do_not_process.’ There can be a
multitude of reasons why one would not want to process certain datasets.

4. Changed $MAGIC_INDEX to just $INDEX.

5. Removed option for building query table so that the user always has
to provide a query table.

6. Changed variable name $queries, which stores the number of queries
found, to $num_queries.

7. Deleted the update history log. This is redundant with what’s on
github and differences between the two logs can be confusing. If we
need to rieinstate it later, I’ve copied and pasted it at the end of
this commit. Please use github to document updates.

8. Removed tabs between LHS and RHS of equal signs. This is a little
difficult to see by eye, and the tabs can be visually offset depending
on the editor used.

9. Implemented basic parameter checks and echo. Reconfigured some
option codes.

######### UPDATE HISTORY
######################################################
# (20-04-15) v0.11 - added option to use an already existing query
table.
#
#                    added subroutine to identify reference genotypes.
if the
#                    keywords for one of the 19 MAGIC parents (e.g.
"edi", case
#                    insensitive) is identified in column "library
name", the
#                    respective reference matrix will be used. if no
genotype
#                    can be identified, col reference will be used.
#
# (21-04-15) v0.12 - Added comments to code.
#
#                    Changed $experiment variable to
#                    $run because 'experiments' are SRX... and can be
shared
#                    across samples. 'Runs' however are unique SRR or
ERR IDs.
#
#                    Changed chdir calls from $cwd/$out to simply $out.
I did
#                    this because if out is a relative path, it will be
created
#                    as such (no need to specify the working
directory). If
#                    its a full path, then having $cwd/$out doesn't
make sense.
#
#                    Kept --split-files on for all fastq-dump calls.
#
#		             Added Sailfish call
#
#                    added checkpoint for successful sailfish execution
based
#                    on the composition of the sailfish logfile.
#
# (22-04-15) v0.13 - changed checkpoint 1: "check if run is already
processed"
#                    into a subroutine.
#
#                    added checkpoint 2: "parse sailfish log".
#
#                    re-arranged data transfer. downloading .sra files
will
#                    now be managed by the slave script fishhook.pl.
#
# (14-05-15) v0.14 - fixed various bugs as descried on github.
#                    added option -p. accepts a path to a file
containing SRA
#                    identifiers that have been processed in a previous
run.
#                    prior to processing a run, checks if this SRA
identifier
#                    is among this list. if yes, skips run.
########################################################################
#######