random subsample
Modified subsample_fastq.pl to sample probabilistically. This allows
one to specify a proportion of reads to keep rather than specifying a
fixed rate of every nth read. With the latter, it’s not clear how to
down sample a sample with 46,000,000 to 40,000,000 reads. But with this
new modification we can specify the frequency parameter as 40/46 and a
random 100*40/46 percent of the reads will be kept.