Wrote code for prospective performance evaluation
Adding evaluate_prospective_performance.m.

This function splits submissions (by date) in to training and test
sets. 90% data is used for training and these are all samples before a
certain date. Remaining 10% of data is used for test.

User supplies expression matrix, query table, and save file name.
1. Expression matrix is OMP decomposed
2. Selected markers are linked to rest of transcriptome using trattoria
3. Tradiction of test set using learned markers and their relationships
to the rest of the transcriptome.

Progress is saved after each step in saveFile.